+++ src/main/java/com/abcdedu_backend/memberv2/adapter/in/dto/request/ChangeMemberRoleRequest.java
@@ -1,4 +1,4 @@
-package com.abcdedu_backend.member.dto.request;
+package com.abcdedu_backend.memberv2.adapter.in.dto.request;
 
 public record ChangeMemberRoleRequest (
         Long memberId
