+++ src/test/java/com/abcdedu_backend/member/service/AdminMemberServiceTest.java
@@ -3,7 +3,8 @@ import com.abcdedu_backend.exception.ApplicationException;
 import com.abcdedu_backend.member.entity.Member;
 import com.abcdedu_backend.member.entity.MemberRole;
 import com.abcdedu_backend.member.repository.AdminMemberRepository;
-import com.abcdedu_backend.member.dto.request.ChangeMemberRoleRequest;
+import com.abcdedu_backend.memberv2.adapter.in.dto.request.ChangeMemberRoleRequest;
+import com.abcdedu_backend.memberv2.application.AdminMemberService;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
